cluster: Add `Cluster` and `Node`
Initial setup of `cluster` module. This module houses the API that will
be used to parse and deserialize cluster configuration data.

Currently, the `Cluster` and `Node` types have been defined in order to
represent the deserializable structures for a given cluster
configuration.

The current idea is to define a cluster that uses three repository
types:

- normal: Regular Git repository.
- bare-alias: Bare Git repository that uses directory as worktree alias.
- root: Special bare-alias repository that houses cluster configuration.

The root (top-level) table of a configuration file will be used to house
the settings for the root repository. While the special "node" table
will be used to house repository entries in the cluster.

The `Cluster` type sets up the `Node` type as a hash table in order to
create a adjacency list such that `Node::depends` will be used to list
edges in the graph. This should hopefully make it easy to implement the
required graph algorithms needed to manage cluster data.

Signed-off-by: Jason Pena <jasonpena@awkless.com>