cluster: Add `Cluster::dependency_iter`
The `Cluster::dependency_iter` method allows caller to produce a path
through the dependencies of a target node in cluster. This dependency
path is determined by `Node::depends` member, which is optional.
Thus, nodes can be defined without dependencies.

In the case that a node has no dependencies, then the path only contains
the node itself and nothing else.

The `DependencyIter` type performs the traversal through the
`Node::depends` field based on the cluter of nodes in `Cluster::node`
member. The `DependencyIter` uses the stack variant of DFS algorithm to
determine the full path of dependencies. This type makes the assumption
that `Cluster` is acyclic.

Test plan:

- Verify that `Cluster::dependcy_iter` passes following cases:
  1. Produces path through node with dependencies.
  2. Produces no path for node without dependencies.

Signed-off-by: Jason Pena <jasonpena@awkless.com>