vcs: Add `syscall_git`
Add `syscall_git` to make system calls to user's Git binary when needed.
Caller can pass in a Git command with a set of arguments, and have the
target repository be treated as normal or bare-alias.

A bare-alias repository is just a bare Git repository that uses the
`--work-tree` flag of the Git binary to force the target repository to
treat a target directory like an alias of a worktree.

Signed-off-by: Jason Pena <jasonpena@awkless.com>