cargo: Add `git2` and `run_script` crates
I added the `run_script` and `git2` crates. The `run_script` crate will
allow the codebase to run shell scripts written by the user for the hook
system. The `git2` library will be used to reduce system calls to user's
binary to speed up execution by directly interacting with the Git
database of a given repository.

Signed-off-by: Jason Pena <jasonpena@awkless.com>