vcs: Redesign for async repo cloning
Massive redesign of the `vcs` module here. Basically, I moved repository
management into two basic types: `RootRepo` and `NodeRepo`. The
`RootRepo` type manages the root repository of a cluster. The `NodeRepo`
type manages the node repositories of a cluster. The special
`NodeMultiClone` type exists to perform multiple async clones of node
repositories in a cluster.

Now the `vcs` module can perform async cloning with visual progress bars
when needed. Although, cloning is a bit slow due to the git2-rs crate
not being the fastest. However, I will tolerate this until I figure out
how to use gix.

The basic git functions I implemented were way to general for the needs
of the project, and did not work well in an async environment. Thus,
this massive redesign was carried out.

I added the `RepoKind::Bare` type to represent pure bare repositories.
Not sure if this is needed, but was useful for mapping out the current
design of the code.

I also added `SparseManip` to make it easier to manipulate sparse
checkout for bare-alias repositories.

Signed-off-by: Jason Pena <jasonpena@awkless.com>