repo: Only set special config settings for bare-alias nodes
Modify `GitWrapper::clone_with_progress` to add the
`status.showUntrackedFiles` and `core.sparseCheckout` settings in the
repositories config file if and only if the repository was defined to be
bare-alias or bare.

This was done, because the user is unlikely to need these settings for
normal repositories at all. Thus, we can reduce interactions with IO by
limiting the writing of these settings to bare-alias or bare
repositories only, because those repositories actually need them.

Signed-off-by: Jason Pena <jasonpena@awkless.com>