cluster: Return key value pair per dependency iter
Turns out I also need the name of a given node in the cluster, which not
stored directly in `Node` but rather as the key to the node in
`Cluster::node` field.

Thus, I modified `DependnecyIter` next to now return `(&str, &Node)` in
order to get `NodeRepo::from_node` to finally work.

Signed-off-by: Jason Pena <jasonpena@awkless.com>