repo: Add `NodeRepo`
The `NodeRepo` type is not responsible for handling operations for node
repository entries in cluster. I also simplified the styling of progress
bars by simply defining the style inside of
`GitWrapper::clone_with_progress` itself rather than having the caller
define the style themselves.

I also updated the log in `GitWrapper::deploy` to now clone a target
repository itself if and only if it does not currently exist for
whatever reason.

Signed-off-by: Jason Pena <jasonpena@awkless.com>