bin: Define basic functionality of deploy command
The deploy command is now implemented (well at least the baseline
version of it). This command takes a list of node names and deploys them
to whatever worktree alias they were defined to use.

It is very bare bones right now. But I just want to implement the clone
command to get a sense of how it should work for the user.

Signed-off-by: Jason Pena <jasonpena@awkless.com>