config: Catch non-existent node in `Cluster::dependency_iter`
I handled the edge-case of the user passing in non-existent node entries
to `Cluster::dependency_iter` by just panicking. This behavior looks
rather ugly and ungraceful.

Thus, I just wrapped the return value in a `Result` and error out if the
user tries to pass in a non-existent node name. This fixes the overall
problem in a fairly simple manner.

Signed-off-by: Jason Pena <jasonpena@awkless.com>