bin: Add `Undeploy` command
Implemented basic skeleton of the undeploy command. I also came to the
conclusion that both the `deploy` and `undeploy` command should not be
allowed to directly operate on the root of the cluster. Undeployment of
the root cluster really does not make sense, because OCD is mainly
designed to function with a root already deployed somewhere.

Signed-off-by: Jason Pena <jasonpena@awkless.com>