bin: Fix non-existent node edge-case in git shortcut
The Git shortcut command that OCD offers does not check if a given node
actually exists in a cluster. This result in OCD panicking when given a
non-existent node to process through the user's Git binary.

This has been fixed by using the `Cluster::get_node` method, _before_
calling Git on a given node to handle this edge-case.

Signed-off-by: Jason Pena <jasonpena@awkless.com>