config: Cleanup cluster module
Added missing documentation, and moved unit tests back into the module
itself. These tests are not that big, and currently it makes sense to do
so.

I also fixed the `smoke_cluster_dependency_iter` test to not fail to
compile anymore with recent changes to the `Cluster::dependency_iter`
method's interface.

Signed-off-by: Jason Pena <jasonpena@awkless.com>