chore(cargo): add expected dependencies
Add a set of dependencies that are expected to be needed for the OCD
project's future development. I also needed to update the deny list to
handle duplicate dependencies and additional licenses that needed to be
allowed in order for the `cargo deny check` command to pass.

I am also moving the project to GitHub to experiment with GitHub actions
and docker CI/CD workflows.

Signed-off-by: Jason Pena <jasonpena@awkless.com>