feat(cluster): add `Cluster` type
The `Cluster` type will be used to provide various operations needed to
manipulate and manage a cluster definition parsed from a configuration
file.

Will be using the `toml_edit::DocumentMut` type to implement the
`Cluster` type as a format preserving TOML parser. This will come in
handy when I begin implementing the commands that need to modify the
user's cluster configuration file in the future.

Signed-off-by: Jason Pena <jasonpena@awkless.com>