feat(cluster): add `Root` type
The `Root` type represents the root of a cluster definition. It is meant
to contain the cluster definition itself so it can later be cloned with
all of the nodes defined for that cluster.

The user can control where the root gets deployed through the
`Root::worktree` field, and can exclude certain files from deployment
using the `Root::excludes` field.

Signed-off-by: Jason Pena <jasonpena@awkless.com>