feat(cluster): add `Node` type to deserialize into hashmap
The `Node` type represents a regular entry in the cluster definition.
Entries are referred to as nodes. A node can either be normal or
bare-alias unlike root being only bare-alias. Nodes can also specify a
URL to directly clone them from when needed. Nodes also allow the user
to specify a target worktree and a set of files they want to exclude
from deployment similar to root. However, what makes nodes really
special is the "depends" field.

A node can specify another node in the cluster as a dependency. Thus,
when the user deploys a node, they can tell OCD to deploy a set of
dependencies for a given node so they do not need to do so manually
themselves!

In the `Cluster::from_str` method, all nodes are extracted into a
hashmap. Root is also extracted as well. This hashmap of nodes uses the
name of a given node as the key, and actual contents of the node as the
value. The hashmap will prove useful for performing DFS and other graph
algorithms in order to implement the dependency deployment system
mentioned earlier.

Finally, I replaced the test case for `Cluster::get_root` with a test
for `Cluster::from_str`. The `Cluster::get_root` routine has been
refactored into a basic getter. However, the `Cluster::from_str` method
is way more complicated now, and makes more sense to properly test now.

Signed-off-by: Jason Pena <jasonpena@awkless.com>