feat(cluster): add acyclic graph check for nodes
In order to obtain the dependencies of a node in a cluster definition,
the set of nodes and their dependencies must be acyclic. I have
configured the node hashmap to represent a directed acylic graph (DAG),
but the user can enter nodes and dependencies however they want when
writing a cluster. To ensure that the graph remains acyclic, a basic
check has now been coded into `Cluster::from_str`.

When a cycle is detected by `Cluster::from_str` a list of nodes causing
the cycle is returned as the error report. The user is expected fix the
cycle(s) themselves.

The acyclic check uses Kahn's algorithm. I find it to be simple to
implement, and rather easy to collect all the problematic nodes to issue
an error report.

Added a basic test to ensure that the acyclic check actually works as
expected for a set of basic cases.

Signed-off-by: Jason Pena <jasonpena@awkless.com>