feat(cluster): add `Cluster::dependency_iter`
Caller can now obtain and iterate through all dependencies of a target
node in a cluster definition. The `Cluster::dependency_iter` method uses
a stack variant of DFS to obtain all dependencies of a target node
through the `DependencyIter` type.

The `Cluster::dependency_iter` method issues a basic log error in case a
given node does not exist in the dependency list for whatever reason,
but keeps iterating until the end.

Test case was added to ensure that `Cluster::dependency_iter` functions
correctly for a node with dependencies, and without dependencies.

Signed-off-by: Jason Pena <jasonpena@awkless.com>