feat(cluster): add `Cluster::get_node`
Caller can now obtain a singular node without its dependencies through
the `Cluster::get_node` method. This method will fail if the node does
not exist in cluster.

Added basic test case to ensure that `Cluster::get_node` works for a
node that exists and a node that does not.

Signed-off-by: Jason Pena <jasonpena@awkless.com>