feat(cluster): add `Cluster::add_node`
Caller can now add a new node into cluster definition using the
`Cluster::add_node` routine. This method will add the new node returning
None if it did not already exist, or Some containing the original node
that it replaced. This method will also insert a new node table if it
does not already exist to add the new node entry into. It can only
really fail if the user already as an item named node, but did not
define it as a table.

Added in test case the ensures that `Cluster::add_node` inserts into an
existing node table, and creates a new node table when it does not
exist.

Signed-off-by: Jason Pena <jasonpena@awkless.com>