feat(fs): add `DirLayout` type
The `fs` module is responsible for providing routines and types that
allow the caller to control the user's file system. The `DirLayout` type
determines the absolute paths to the most important directories that OCD
needs to have access to. Currently `DirLayout` locates paths for the
user's home directory, the `$XDG_CONFIG_HOME/ocd` directory for
configuration files, and the `$XDG_DATA_HOME/ocd` directory to house
repository data.

Signed-off-by: Jason Pena <jasonpena@awkless.com>