feat(vcs): add `Git::clone_with_progress`
Caller can now perform a Git clone without needing to use
`Git::bincall` via `Git::clone_with_progress`. The cool thing about this
method is that it provides a progress bar to show the progress of the
clone itself.

This clone method also performs authentication when needed through the
auth_git2 crate. I added a custom prompter for terminal authentication
that will block either single or multi progress bars to prevent the
interleaving and over-writing of user input with progress bar drawing.

The current plan is the use the `Git::clone_with_progress` to perform
async cloning of multiple node repositories in the future.

I also added some setter and getter methods to make it easy to obtain
field data from the `Git` type when needed. I also added `Git::new` to
construct a new `Git` type with the help of `fs::DirLayout`.

Signed-off-by: Jason Pena <jasonpena@awkless.com>