feat(vcs): allow `NodeRepo` and `RootRepo` to make calls to git
This commit prepares the `vcs` module to provide functionality required
for the Git command shortcut offered by OCD's CLI. The user is allowed
to make direct calls to their Git binary for target repositories in
their cluster through OCD.

The `gitcall` routine is what will be used. I also added helpers like
`RootRepo::from_cluster` to make it easier to create a root repository
when it is already known that it was successfully deployed.

Signed-off-by: Jason Pena <jasonpena@awkless.com>