feat(vcs): add repository deployment functionality
Initially the `RootRepo` type had a feature to deploy the root
repository after cloning it through `RootRepo::deploy`. However, in
order to implement the deploy command for the CLI, `NodeRepo` type needs
to have its own version of `deploy` method.

I also gave `SparseManip` the ability to include the full index of a
repository in sparsity rules in order to allow the user to fully deploy
a given node or root with excluded files when the want. This
functionality is reflected by the `deploy_all` methods that were added.

Signed-off-by: Jason Pena <jasonpena@awkless.com>