refactor(bin): convert to lib+bin crate
I want to perform integration testing on certain components of the
codebase. However, these tests will be interacting with the file system,
which would slow down unit testing. Thus, I want to move those
integration tests to the `tests/` folder at the top-level of the
project.

However, I cannot access the internal API used by the project given that
it is only a binary. Thus, I decided to convert the codebase to a
`lib+bin` crate, where the library portion of the project contains the
internal API that I need access to for integration testing.

I also enabled a few clippy lint permissions regarding documentation for
the project in the `src/lib.rs` file itself, because I plan on cleaning
up the documentation of the codebase as well.

Signed-off-by: Jason Pena <jasonpena@awkless.com>