feat(utils): add `syscall_interactive`
The `syscall_interactive` function will be used to allow the caller to
make calling shell programs more interactive for the user. The idea is
to use this method for the Git shortcut command, so the user can utilize
their Git program more interactively.

I also added a integration test to make sure that it works as expected,
similar to the test for `syscall_non_interactive`.

Signed-off-by: Jason Pena <jasonpena@awkless.com>