Add `README.md` file
I finally setup the `README.md` file to introduce newcomers to the OCD
project. There are a few sections that have not been filled out yet,
which have been labeled with a _TODO_. I am still figuring out how to
properly setup these unfinished sections, which will be filled out as
development moves forward. For now, the `README.md` file provides a
basic summary of the project, states it is open to contribution, and
provides information about its copyright and licensing. This should be a
good enough start.

Signed-off-by: Jason Pena <jasonpena@awkless.com>