Add code quality checks as a CI workflow
Now whenever a pull request or push to `main` branch is done, the code
quality check CI workflow will be activated through GitHub Actions. This
will be used as the base step for code review of changes to the codebase
of the OCD project.

For now the quality check includes:

- Unit and integration testing checks through Linux and Windows.
- Code formatting check.
- Code linting check.
- REUSE 3.3 compliance check.
- Valid dependency license check through `cargo-deny`.

Signed-off-by: Jason Pena <jasonpena@awkless.com>