model: Add acyclic check to `Cluster::from_str`
Now whenever a call is made to `Cluster::from_str` the
`Cluster::acyclic_check` method will be run to ensure that node
dependencies in the cluster are acyclic.

The algorithm I decided to use to for `Cluster::acyclic_check` is Kahn's
algorithm. It is very simple, and easy to test.

Test plan:

- Smoke test `Cluster::from_str` acyclic check:
  1. Does not fail for single node with no dependencies.
  2. Does not fail for a set of nodes with acyclic dependencies.
  3. Catches a node that depends on itself.
  4. Catches a full cycle in a set of nodes.

Signed-off-by: Jason Pena <jasonpena@awkless.com>