model: Fix broken tests
Now that `config_dir` and `data_dir` create their paths in full, they
try to create the fake paths for the smoke tests in `model` for the
`Cluster` type. This leads to failed test not having the permission to
create stuff in `/` because I was dumb enough to have fake directories
point to that path.

Signed-off-by: Jason Pena <jasonpena@awkless.com>