bin: Implement deploy command
User can now deploy node repositories to their expected directory
aliases through the new deploy command! This command will take a set of
patterns, match them to the entries in their cluster definition, and
deploy the matched entries. The user can specify if they want
dependencies of matched entries to be deployed or not. They can even
deploy the entire index of of an entry with excluded files.

The root repository is always deployed, so this command does not do much
other than allow the user to deploy the full index of root when they
want.

Signed-off-by: Jason Pena <jasonpena@awkless.com>