store: Add remove command
User can now fully remove both nodes and root from their cluster.
Removing root will nuke the entire cluster from existence. This command
will also ensure that the target repository is fully undeployed before
removing it from the cluster definition and the repository store.

Just learned that libgit2 needs an initial commit in order for it to
function properly. I currently do not know if this is an issue with how
I am using libgit2, or if libgit2 has not yet fixed this issue.

Signed-off-by: Jason Pena <jasonpena@awkless.com>