store: Fix empty repository check
Originally, the `Git::is_deployed` method checks for empty index as its
core logic for checking an empty repository. Turns out this logic is
broken. I mentioned in a previous commit about the unborn branch error I
was encountering. This error occurred because I was not checkng for an
empty repository correctly.

The `Git::is_empty` method was added to amend this issue. It does not
check for an empty index, but instead checks for a repository with no
commits in its tree. The idea of checking for an empty index was rather
dumb on my part, yet it baffles me on how most of the code was still
working for many common cases during manual testing. I really need to
review my manual test setup, plus I really need to convert this setup
into Rust code somehow.

Plus, I also fixed the file matching logic in `Git::is_deployed`,
because it also relied on the index to work. Luckily, Eric's code helped
with this issue[1]. Now the code revwalks through the tree pointed to by
HEAD as a commit, and extracts file names for each entry of the obtained
tree. This works way better.

Finally, I added the `gitcall` method to `Root` and `Node` to allow for
interactive calls to Git through these types for the new git shortcut
command of OCD's CLI.

[1]: https://www.hydrogen18.com/blog/list-all-files-git-repo-pygit2.html

Signed-off-by: Jason Pena <jasonpena@awkless.com>