store: Add `TablizeCluster` type
Now we can tablize the user's cluster in order to obtain some important
information at a glance for the list command.

The `TablizeCluster::names_only` only lists the names of all
repositories in the repository store. The `TablizeCluster::fancy` shows
the deployment type, entry name, deployment status, and current branch
of each entry in repository store.

The magic of the tablized data comes from the `tabled` crate.

Signed-off-by: Jason Pena <jasonpena@awkless.com>