Fill in TODOs in `README.md`
I now know what information I need for the installation and usage
sections of the `README.md` file as development towards the first
release draws near. I think some more information is needed for what you
can do with OCD in the usage section, but this is a good start for now.

Signed-off-by: Jason Pena <jasonpena@awkless.com>