bin: Fix failure to deploy without entry in repository store
I made a stupid mistake of not using the name of the dependency as the
check for whether it exists in the repository store or not. Instead I
used the name of the target node causing the check to always pass even
when the dependency has not entry in repository store.

Signed-off-by: Jason Pena <jasonpena@awkless.com>