store: Refactor `Root` type
This commit is the first step in refactoring the `store` module. I
noticed that the `Git` type was slowly becoming a god object. I wanted
to decompose it into smaller self contained types. This commit does
that.

It is a little disorganized, but a set of new types have been introduced
to replace the `Git` type:

- `RepoEntry` contains routines to manipulate a repository entry in the
  repository store.
- `RepoEntryBuilder` is a builder for `RepoEntry` similar to
  `GitBuilder` for `Git`.
- `Deployment` is a trait that offers to deployment strategies for
  repository store.
- `NormalDeployment` is a strategy for normal repository deployment.
- `BareAliasDeployment` is a strategy for bare-alias repository
  deployment.
- `RootDeployment` is a strategy for root repository deployment.

All of these types are meant to replace `Git` type in full. For now I
used these types to simplify and retain the functionality that the
`Root` type offers compared to when it was just a wrapper over `Git`
type.

I also made sure that file extraction from repository tree was
simplified and made exclusive to `Root` itself in order to obtain its
cluster. This new exclusive method is `Root::extract_cluster_file`. Only
root needs to do this, node repositories do not.

All in all, this is the first step of replacing `Git` with the new
modular types. The `Root` type was first to no longer use `Git` type.
Now the `Node`, and `MultiNodeClone` types are up next. Once these are
done, I can finally remove the `Git` type in full.

Signed-off-by: Jason Pena <jasonpena@awkless.com>