fs: Make `config_dir` create configuration directory
Add in fail-safe to `config_dir` to ensure that the configuration
directory exists no matter who calls it. This also removes the duplicate
code of configuration directory creation from `load` and `save` into one
place.

This also fixes a bug with the clone command where it will fail to
deploy root because the configuration directory does not exist.

Signed-off-by: Jason Pena <jasonpena@awkless.com>