tests: Clean up testing of `store` module
With all the new refactored code, new tests will need to be developed.
Currently, I am manually testing the code. I will convert my manual
setup into an automated test setup in Rust, but for now, the old tests
are broken, so they need to be fixed and cleaned up.

Signed-off-by: Jason Pena <jasonpena@awkless.com>