store: Fix locating cluster definition in root
Found basic bug where using the home directory as a directory alias for
the root repository would not locate the cluster definition correctly.
I forgot that libgit2 treats directories as trees. Thus, using a regular
path to locate a file in a repository only works if you iterate through
each component of the path as a tree. I did not do this, so the
`Root::extract_cluster_file` method would always fail to find a cluster
definition in `.config/ocd/cluster.toml`, because I never walked the
tree to reach that path component.

This is now fixed, with the corresponding test cases being updated to
more accurately test this behavior.

Signed-off-by: Jason Pena <jasonpena@awkless.com>