model: Add hook system
The `model` module contains the full implementation of the hook system.
The hook system allows the user to execute hook scripts for a given OCD
command.

The `HookRunner` type is not format preserving, because it will never
modify the special hooks configuration file that the user defines. It
only needs to read, deserialize, and read and execute a given hook
script entry.

Hook entries can be executed _before_ or _after_ a given command, and
can be set to only execute for a given target entry in the cluster
definition. One of the coolest features I've implemented is the prompt
feature. `HookRunner` can be modified to perform three actions with hook
entries for a given command: `HookAction::Always`, `HookAction::Never`,
`HookAction::Prompt`. The `HookAction::Always` variant will always
execute hook entries no questions asked. The `HookAction::Never` variant
will never execute hook entries no questions asked. The
`HookAction::Prompt` variant will make `HookRunner` execute a pager to
present the contents of the hook in question to the user. The user can
view the hook and either it the "a" key to accept, or the "d" key to
deny the hook from executing. This will be done for all hook entries
defined.

Signed-off-by: Jason Pena <jasonpena@awkless.com>