model: Add `Cluster::new`
The `Cluster::new` method allows caller to open, read, and deserialize
cluster configuration files into the respective `RootEntry` and
`NodeEntry` types. This method ensures that there is exists one root for
any new `Cluster`, and collects all node entries into a hashmap in order
to represent them as a DAG.

The new testing framework is something I find really cool. Much like the
way in which I began testing repository logic in the `store` module, I
am using the txtar format to perform tests on the `Cluster::new` method.
Each fixture is written in txtar format, and based into a given test
function as a test case. The contents of the fixture is extracted, and
written to wherever I need it to be for the test to work. This greatly
simplifies the setup of each test for `Cluster::new` and allows my to
easily test various edge cases without having to modify the test code
itself to much.

Signed-off-by: Jason Pena <jasonpena@awkless.com>