Remove error variant logic
I am going to stick with just `anyhow` crate code instead of using
custom error type code. Using custom error types only makes sense if you
intend on distributing the code as a library. However, OCD is not a
library, it is a binary. Thus, it makes more sense to stick with just
anyhow for error handling.

Signed-off-by: Jason Pena <jasonpena@awkless.com>