store: Remove `Node::new_clone`
This old method is no longer needed anymore, because `Node::new_open`
performs the cloning of a node if it does not exist, and the
`MultiNodeClone` type is expected to be used most of the time when it
comes to cloning nodes in repository store. Thus, `Node::new_clone` is
redundant.

Signed-off-by: Jason Pena <jasonpena@awkless.com>