store: Fix root exclude file deployment
Forgot to use `RepoEntryDeployer::add_excluded` to actually add the
excluded files from the extract root configuration file settings. This
causes `Root::new_clone` and `Root::new_open` to deploy the entire
repository to the work-dir alias instead of excluding certain files as
expected.

I also had to fix test cases for this, because this fixed caused some of
them to fail.

Signed-off-by: Jason Pena <jasonpena@awkless.com>